# [Fix] Increase speed of transform_points