### Description
The speed of the function `transform_points` in `linalg.py` is increased by using `torch.bmm` instead of `torch.matmul` as done in the corresponding function in Pytorch3d. As `torch.matmul` handles different shapes via broadcasting but `torch.bmm` does not, we need to include extra code to reshape the data. The increase in speed for some input sizes is:
Before:
```
Benchmarking project_points
inp=(64, 1048576, 3), dev=cuda:0, 0.6896541118621826, sec
inp=(8192, 8192, 3), dev=cuda:0, 0.49063897132873535, sec
inp=(1048576, 64, 3), dev=cuda:0, 0.5144214630126953, sec
```

After modification:
```
Benchmarking project_points
inp=(64, 1048576, 3), dev=cuda:0, 0.27915072441101074, sec
inp=(8192, 8192, 3), dev=cuda:0, 0.10534358024597168, sec
inp=(1048576, 64, 3), dev=cuda:0, 0.11871767044067383, sec
```
However, even though the new function passes all the tests, the results are not exactly the same as the original function. The new function matches the results given by Pytorch3d. This Colab notebook shows a quick example of that:
https://colab.research.google.com/drive/1dVHXicxHVPfUdyuYNdd4jOzx7JKtwqxb#scrollTo=ixEUsV33ofYE

Changes:
 - `transform_points` in `linalg.py` is changed to use `torch.bmm` instead of `torch.matmul`. Also added code to reshape data.
 - Added a new performance test, `test_project_points_speed.py`, to test the performance of `transform_points`.
 - Small non-related changes in `test_augmentation_mix.py` to fix some cuda tests crashing. Some `device` variables were missing.

### Status
Ready

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not applicable items -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or new feature that would cause existing functionality to change)
- [ ] New tests added to cover the changes
- [ ] Docstrings/Documentation updated


## PR Checklist
### PR Implementer
This is a small checklist for the implementation details of this PR.

If there are any questions regarding code style or other conventions check out our 
[summary](https://github.com/kornia/kornia/blob/master/CONTRIBUTING.rst).

- [ ] Did you discuss the functionality or any breaking changes before ?
- [X] **Pass all tests**: did you test in local ? `make test`
- [ ] Unittests: did you add tests for your new functionality ?
- [X] Documentations: did you build documentation ? `make build-docs`
- [X] Implementation: is your code well commented and follow conventions ? `make lint`
- [X] Docstrings & Typing: has your code documentation and typing ? `make mypy`
- [ ] Update notebooks & documentation if necessary

### KorniaTeam
<details>
  <summary>KorniaTeam workflow</summary>
  
  - [ ] Assign correct label
  - [ ] Assign PR to a reviewer
  - [ ] Does this PR close an Issue? (add `closes #IssueNumber` at the bottom if 
        not already in description)

</details>

### Reviewer
<details>
  <summary>Reviewer workflow</summary>

  - [ ] Do all tests pass? (Unittests, Typing, Linting, Documentation, Environment)
  - [ ] Does the implementation follow `kornia` design conventions?
  - [ ] Is the documentation complete enough ?
  - [ ] Are the tests covering simple and corner cases ?
 
</details>